backend/requirements.txt
openai==1.14.3
python-dotenv==1.0.1
fastapi==0.110.2
uvicorn==0.29.0
pydantic==2.7.1
requests==2.31.0
httpx==0.27.0
sqlalchemy==2.0.29
asyncpg==0.29.0
jinja2==3.1.3
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
# ai_assistant - локальный модуль, не требуется в requirements.txt


